+++ here-naksha-lib-core/src/test/java/com/here/naksha/lib/core/models/storage/StringCodec.java
@@ -30,7 +30,7 @@ import org.jetbrains.annotations.NotNull;
 class StringCodec extends FeatureCodec<String, StringCodec> {
 
   @Override
-  protected Short getDefaultGeometryEncoding() {
+  protected Integer getDefaultGeometryEncoding() {
     return GEO_TYPE_WKB;
   }
